ml/requirements.txt
# Core ML Libraries
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0

# SAM2 Dependencies (SAM2 installed separately via Makefile)
transformers>=4.35.0
hydra-core>=1.3.0
iopath>=0.1.10

# API & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# HTTP Client (using httpx as requested)
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0

# Image Processing
albumentations>=1.3.1

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2

# Distributed Training
accelerate>=0.24.0

# Database for USDA lookups
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Development
pytest>=7.4.0
black>=23.10.0
isort>=5.12.0
